Refactoring: additional level for connectors
we want to have it this way: to include only the source, nothing more. All logic for connector selection should be within source package
