+++ source/connectors/github/helpers.go
