+++ source/connectors/github/internal/client/interfaces.go
