+++ source/connectors/github/internal/testclient/testclient.go
@@ -23,7 +23,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/artem-sidorenko/chagen/connectors/github/internal/client"
+	"github.com/artem-sidorenko/chagen/source/connectors/github/internal/client"
+
 	"github.com/google/go-github/github"
 )
