+++ source/connectors/interface.go
